Tweak model to facilitate traversals for rentals and repairs; improve query for rentals and repairs

+++ HW1/src/main/java/com/example/demo/models/PlantsWithRentalsAndRepairs.java
@@ -12,8 +12,8 @@ public class PlantsWithRentalsAndRepairs {
 
 	PlantInventoryEntry entry;
 
-	Integer rentals;
+	Long rentals;
 
-	Integer repairs;
+	Long repairs;
 
 }
